Remove support for Go1.6-1.8
