+++ benchmark/benchmark_test.go
@@ -1,5 +1,3 @@
-// +build go1.7
-
 /*
  *
  * Copyright 2017 gRPC authors.
