+++ credentials/syscall.go
@@ -1,4 +1,4 @@
-// +build !linux !go1.9 appengine
+// +build !appengine
 
 /*
  *
@@ -18,9 +18,18 @@
  *
  */
 
-package channelz
+package credentials
 
-// GetSocketOption gets the socket option info of the conn.
-func GetSocketOption(c interface{}) *SocketOptionData {
-	return nil
+import (
+	"errors"
+	"syscall"
+)
+
+// implements the syscall.Conn interface
+func (c tlsConn) SyscallConn() (syscall.RawConn, error) {
+	conn, ok := c.rawConn.(syscall.Conn)
+	if !ok {
+		return nil, errors.New("RawConn does not implement syscall.Conn")
+	}
+	return conn.SyscallConn()
 }
