+++ encoding/proto/proto_benchmark_test.go
@@ -1,5 +1,3 @@
-// +build go1.7
-
 /*
  *
  * Copyright 2014 gRPC authors.
