+++ internal/syscall/syscall_linux.go
@@ -1,4 +1,4 @@
-// +build !appengine,go1.7
+// +build !appengine
 
 /*
  *
